+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added tag classes for group E.7.1.2
 -   Added tag classes for group E.7.2
 -   Added tag classes for group E.7.2.1
+-   Added tag classes for group E.8
 -   Base classes of the Builder pattern are added for the generation of Factura Electronica
 -   Factura Electronica generation test added